+++ ESP/Visual/Heightmap.cs
@@ -6,6 +6,28 @@ namespace ESP;
 public partial class Visual
 {
   private static readonly List<GameObject> renderers = new();
+  // Copy paste from old Valheim code.
+  private static Vector3 CalcNormal(Heightmap hm, int x, int y)
+  {
+    Vector3 vector3_1 = hm.CalcVertex(x, y);
+    Vector3 rhs;
+    if (x == hm.m_width)
+    {
+      Vector3 vector3_2 = hm.CalcVertex(x - 1, y);
+      rhs = vector3_1 - vector3_2;
+    }
+    else
+      rhs = hm.CalcVertex(x + 1, y) - vector3_1;
+    Vector3 lhs;
+    if (y == hm.m_width)
+    {
+      Vector3 vector3_3 = hm.CalcVertex(x, y - 1);
+      lhs = vector3_1 - vector3_3;
+    }
+    else
+      lhs = hm.CalcVertex(x, y + 1) - vector3_1;
+    return Vector3.Cross(lhs, rhs).normalized;
+  }
   public static void DrawHeightmap(Vector3 pos, float radius)
   {
     List<Heightmap> hms = new();
@@ -24,7 +46,7 @@ public partial class Visual
           var vector = hm.transform.position + hm.CalcVertex(j, i);
           var vertex = hm.CalcVertex(j, i);
           if (Utils.DistanceXZ(vector, pos) > radius) continue;
-          var normal = hm.CalcNormal(j, i);
+          var normal = CalcNormal(hm, j, i);
           var line = Visualization.Draw.DrawLineWithFixedRotation(Tag.Terrain, hm, vertex, vertex + normal * 1f);
           var paint = hm.GetPaintMask(j, i);
           line.GetComponent<LineRenderer>().material.SetColor("_Color", paint);